requirements.txt
django
djangorestframework
coverage
pre-commit
psycopg2-binary
django-cors-headers